# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.UserSW:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      position:
        type: string
      role_id:
        type: integer
    type: object
  utils.DataResponse:
    properties:
      error_code:
        type: integer
      message: {}
    type: object
host: localhost:5005
info:
  contact: {}
  description: Api INSIGHT project
  title: Insight API SERVICE
  version: 1.0.0
paths:
...
  /users/by-id:
    get:
      consumes:
      - application/json
      description: Просмотр пользователя
      operationId: getUserById
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getUserById
      tags:
      - Users
  /users/edit:
    put:
      consumes:
      - application/json
      description: Редактирование пользователя
      operationId: editUser
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.UserSW'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editUser
      tags:
      - Users
  /users/list:
    get:
      consumes:
      - application/json
      description: Просмотр списока пользователей
      operationId: getAllUsers
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllUsers
      tags:
      - Users
  /users/me:
    get:
      consumes:
      - application/json
      description: Информация про меня
      operationId: getMe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getMe
      tags:
      - Users
  /users/new:
    post:
      consumes:
      - application/json
      description: Добавление нового пользователя
      operationId: addNewUser
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.UserSW'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addNewUser
      tags:
      - Users
  /users/rm:
    delete:
      consumes:
      - application/json
      description: Удаление пользователя
      operationId: deleteUser
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteUser
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"